Add design system JS dependencies
Stimulus will be used to apply JS sprinkles to view components in a concise,
testable way.

We can leverage highlight.js as a nice syntax highlighter - we could also use
Coderay for this, but the GOV UK design system appears to use highlight.js,
so we should probably be consistent.