Use radius instead of radius_in_km
The API was incorrectly exposing a `radiusInKm` query string attribute as well
as a `radius` attribute when searching for events. The former is a readonly
computed attribute so setting it would have no effect. Updating to use the
`radius` attribute instead will filter correctly.

We were also passing a distance collected in miles to `radiusInKm` which would
not have returned accurate results. The `radius` attribute, however, does
accept miles so this should now be fine.