Add authentication step to event sign up
If a user enters first/last/email that matches a candidate already in the CRM
then they will be taken to a new `Authenticate` wizard step where they will be
asked to enter the verification token that has been emailed to them.

In the event that the user cannot be found or if the CRM is not reachable (the
API throws an exception) then the user will be treated as a new candidate and
the `Authenticate` step will be skipped.

Renames `email_address` to `email` throughout for parity with the API (making
request logic slightly easier).