Verify code on Authenticate step
When a user enters a code we first check if it is syntactically valid (present, numerical and exactly 6 digits). If it is, then the code is sent to the API to check if it is correct.

We only hit the API once per code (so if the code changes, we will re-call the API to verify). This is to avoid multiple calls via `valid?`.

If the code is correct, then the wizard store is updated with the response from the API. The matchback criteria will not be overriden (email, first name, last name).