Send request to add attendee to event
Update the wizard to send a request to the API to add the attendee to an event.

It is assumed that the latest privacy policy is being accepted, however this
will need to be updated to match the actual policy we show to the candidate (we
don't link to the policy yet).

If the request to the API fails a 500 error will currently be thrown; we may
want to handle this more gracefully in the future.