Connect re-send verification link to API
When a user clicks the resend verification link on the `Authenticate` step of
event sign up, we issue another request to the API to generate a TOTP and
email it to the user.

To make the mechanism work without Javascript we issue a GET request and
retrieve the existing candidate details from the wizard store. This way no PII
is retained in query string parameters and the request can be made without
Javascript (which wouldn't be the case if it was a POST).

As part of the request a redirect path is sent to ensure the user returns to
the `Authenticate` step, with an additional message visible:

> We've sent you another email.