Switch to using readable_id of TeachingEvent
We currently use the `id` attribute of the `TeachingEvent` in URLs - this is a
long `Guid` that is not SEO friendly. Instead, the `TeachingEvent` object now
exposes a `readable_id` that is a small integer, followed by the event name
(slugified) and a timestamp.

This commit updates the routes to use the `readable_id` instead of the `id`.