Build navigation menus from front matter
The content repo is being updated so that pages include a `navigation` front
matter attribute containing the position of the page in the header/footer
navigation menus. This commit adds a helper to find pages that should appear in
the navigation and updates the static links to be built out dynamically.