Fix exception when event type can't be found
We have hard coded some event types that don't yet exist (Online events), so
when we lookup their name `nil` is returned - we are then trying to
`parameterize` the `nil` object, which causes an exception.

Defaulting to an empty string fixes this for now; going forward the events will
be connected to the API so this problem will go away.