Monkey patch GetIntoTeachingApiClient::ApiClient with caching
Replaces the Typhoeus HTTP library with Faraday and the `faraday-http-cache`
middleware to ensure `ETag` headers are adhered to.

Adds a blanket 10 minute cache on all GET responses; they will be cached for an
hour under the `ETag` header anyway, but this still requires a round trip to
the server to check - it should be safe to cache locally for 10 minute periods
and this ensures pages that make multiple requests to the API (such as the
event listing page) are still fast. I tried initially to implement this
`Cache-Control` header in the  API, however ASP.NET Core is fairl strict and
appears to ignore `Cache-Control` headers behind authentication at a low level,
making it hard to change.