Show correct sign up method for event type
If the event has a `web_feed_id` present, we can offer the online registration
form.

If the event does not have a `web_feed_id` then it should have either a
`provider_website_url`, in which case we can link out to the website, or a
`provider_contact_email` that we can use to prompt the user to email for
attendence.