Show all events, categorised
Previously, the events landing page displayed the first 3 events from each
category (of the current month) and all the events in the current month at the
bottom. When you performed a search you were always presented with the results,
uncategorised.

Instead, this commit removes the 'all events' section and displays all the
event results, categorised.