Do not display the full address of online events
If an event is online it will still have an associated `building`. We should
not display the full address/venue information to users in case they try and
turn up for the online event in person.

Instead, the city will be displayed for online events in eventbox/short
partials and the venue section will not be displayed as part of the view event
page.