Reset GOV.UK styles that are unintentionally modified
As well as our own stylesheet, we also import the GOV.UK design system styles.

There are a number of generic selectors used in our stylesheets (anchor tags,
lists, etc) that end up adding attributes to some of the GOV.UK styles.  We
need to undo these style changes when they occur, which is what this set of
styles is for.