Include matchback fields in export
When an existing candidate successfully matches back to a contact in the CRM we
populate the store with their existing details, including `candidate_id` and
`qualification_id`. As we scope the information sent back to the API down to
only the attributes of steps that weren't skipped - and
`candidate_id`/`qualification_id` are not attributes of any step - the ids
weren't being sent back to the API.

This adds `candidate_id` and `qualification_id` to the `export` call on the
`Authenticate` step and also ensures no answers to the `Authenticate` step will
ever appear in the review answers view.