Remove sign up/how to attend from closed events
If an event has a status of 'Closed' then it should not be possible to sign up
or enquire about attendence.

This commit hides the 'Sign up for this event' buttons from the event detail
page if the event accepts online registration and also removes the 'How to
attend' section if the event is closed.