Prompt for feedback when closing window
When a user moves their mouse to close the window/tab, we want to prompt them
to leave feedback.

Add `FeedbackPromptController` to intercept when the mouse leaves the top of
the viewport and display feedback dialog partial. Set a cookie when the prompt
has been seen so we only show once per browser session. Set a cookie when the
prompt has been actioned so that we don't show for another 12 months.

Consolidate the cookie acceptance CSS to make it a more generic `dialog.scss`.