Obtain CRSF token via ajax request
Having the CRSF token render out as part of the template breaks the page-level
cache.

Add a new `SessionsController` that contains an action that returns the CRSF
token for the session.

Update `PageHelpfulController` to call out and obtain a token prior to
submitting the feedback.

Use wait-for-expect to reliably test asynchronous code.