Add TTA service to connect_src
Sentry flagged a CSP violation against linking out to the TTA
service. It happens due to the link appearing as an internal
path `/tta_service`, which then redirects to an external URL
(the TTA service host).

Adding `turbolinks: false` to the `redirect_to` should cause a
full page reload, but for some reason Turbolinks is still
performing the XHR request.

Adding the `TTA_SERVICE_URL` host to the `connect_src` CSP
silences the error and should be safe enough as this is one of
our services.