Persist event filters in category back link
When navigating from a filtered set of search results to an event category
we display a back link to 'All events'. Currently, this links to `/events`
and you lose your filters; instead, we want to persist the filters and
display the results the user navigated away from initially.

Use the referrer to point back to the filtered results; if the referrer was
external or not available we fallback to the `/events` path.