Allow symlinking of app/views/content directory
Copying the content directory, as outlined in the README, is a bit of a
chore. It would be nice to have the option of symlinking it directly.

In order to do this the relevant .gitignore line needs to be changed so
the entire directory is ignored[0] and not just its contents.

To create a symlink from the app/views directory do:

$ ln -s /the/path/to/your/content/dir content

Then ensure it worked (I used a relative path):

$ ls -l
total 44
drwxr-xr-x. 2 peter peter 4096 Oct  1 12:16 components
lrwxrwxrwx. 1 peter peter   42 Oct  1 13:34 content -> ../../../get-into-teaching-content/content

[0] https://git-scm.com/docs/gitignore#_pattern_format