Hard code the event category names
We were previously inferring the category names from the CRM types; this was
fine for all of them apart from 'School or University Event', which needs to be
shown as 'School and University Event'. It doesnt make sense to change this in
the CRM since it's referring to the singular and _or_ makes more sense in that
context.