Reapply the changes made to event name lookup
The type_name translation lookup was lost in my rebase because of the
replacement of `_eventbox.html.erb` (ad9802e91bd00) so this reapplies
it.

The use of translation had a minor side effect that when a unexpected
event type is encountered a class name called
`event-resultbox__divider--span-class-translation_missing-title-translation-missing-en-event_types-name-singular-singular-span`
is emitted.

This is ignored for the moment but should be tidied up