Only render the first 9 events per category
Now there's been an influx of online events the index page is a bit
unwieldy. After a discussion, doing this in the view was deemed the
most straightforward approach for the time being.

The capping limit is set via the INDEX_EVENTS_PER_TYPE_CAP constant in
the EventsController and is not applied when there are any filters
applied.