Update the events listing to use new presenter
Instead of traversing a multi-layered hash the categories (get into
teaching, school and university events) are rendered one after the
other. This change amends the listing to make use of the events in the
new format.

The main change is now instead of breaking into groups of 3 and
rendering row by row the entire list is rendered and we'll use CSS to
display it in a grid.