Update the CSS to lay the events out in rows of 3
This uses CSS Grid[0], which is supported in all major browsers. The
fallback is to display the events in a linear list (as it would display
on mobile). This does have the possiblity of alienating IE10 users who
also have high definition displays, we can add a further Flexbox
fallback if necessary.

[0] https://caniuse.com/css-grid