Improve CSS fallback for IE10 and IE11
Even though IE11 does support CSS-Grid, it's an earlier incompatible
version of the spec[0]. This change first styles the list container
using Flexbox and then overrides with Grid where it's supported. Now the
events are displayed in a tiled layout everywhere back to IE9, where
they're displayed in a list (like the mobile view).

[0] https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Grid_Layout/CSS_Grid_and_Progressive_Enhancement