Add a wrapper that builds story using frontmatter
Currently the stories are very HTML-heavy Markdown files. Much of the
HTML is repeated across stories so this is an effort to move as much of
the repeated code as possible to the templates and populate them with
data stored in the frontmatter.

The `build_layout_from_frontmatter` flag allows this approach to be
toggled on a per-story basis. It'll allow us to migrate slowly without
having to have a big bang release.