Reduce topExitSensitivty to 0
With the top sensitivity set to 20 the feedback prompt will appear if the mouse
gets within 20 pixels of the top of the viewport. We feel this is too sensitive
given that it likely overlaps the navigation menu a little and will cause a
false positive if someone moves their cursor over the top edge of a menu item
they are trying to click.

Reduce to 0 so that the cursor has to exit the view port entirely before the
feedback prompt is triggered.