Convert tests to use jest.useFakeTimers()
Previously we were using `setTimeout` within the spec itself to delay before
making an assertion. This is fine for short timeouts, but we are introducing a
longer 20 second timeout to delay the feedback prompt appearing - using jest
fake timers will let us advance the time instantly and avoid slow tests.