Prevent feedback prompt overlaying cookie prompt
When we migrated to `CookiePreferences` this behaviour was lost; in addition to
checking if functional cookies are allowed we also need to ensure `cookieSet`
is `true` (functional cookies are allowed even before the user accepts
cookies). The `cookieSet` flag indicates that the user has manually accepted
cookies and the prompt has disappeared.