Delay feedback prompt for 20 seconds
When the user loads the page if they immediately move their mouse to the top of
the viewport (or scroll up on mobile) the feedback prompt will display.

This doesn't give much time for the user to interact with the website and may
annoy them if they have just landed on a page they didn't want to be on.

Instead, this commit adds a 20 second delay, so the user must view the page for
at least 20 seconds before their behaviour will trigger the feedback prompt to
appear.