Whitelist CSP keys and cap entry length
To help prevent malicious usage of the endpoint we ensure the `csp-report` only
contains valid keys and truncate the length of all values (hopefully avoiding
someone maliciously filling up our logstash).

I thought about setting up the CSP report as a model to add further validation
(ensure the blocked-uri is a URL, for example) but it feels like more hassle
than it's worth.