Strip query params from document/blocked URIs
We are getting a lot of similar CSP violations where the blocked/document URIs
only differ by their query string parameters, for example:

```
https://t.co/i/adsct?type=javascript&version=1.1.0&p_id=Twitter&p_user_id=0&txn_id=o0wo4&events=%5B%5B%22pageview%22%2Cnull%5D%5D&tw_sale_amount=0&tw_order_quantity=0&tw_iframe_status=0&tw_document_href=https%3A%2F%2Fbeta-getintoteaching.education.gov.uk%2Fsteps-to-become-a-teacher

https://t.co/i/adsct?type=javascript&version=1.1.0&p_id=Twitter&p_user_id=0&txn_id=o0wo4&events=%5B%5B%22pageview%22%2Cnull%5D%5D&tw_sale_amount=0&tw_order_quantity=0&tw_iframe_status=0&tw_document_href=https%3A%2F%2Fbeta-getintoteaching.education.gov.uk%2Fsteps-to-become-a-teacher%23step-4
```

If we truncate the query parameters it will let us aggregate on the domain/path
and provide a better insight into the CSP violations. The full violation
details are still available to explore in Kibana.