Retain client IP address in Wizard
We want to be able to access the end user IP address from the wizard step that
makes the request to the `create_candidate_access_token` endpoint. If we pass
this as the `X-Client-IP` header value it will cause the API to rate limit the
endpoint for the individual user, instead of for the app as a whole (based it's
API key).