Transition events#show_category to new search method
We are transitioning away from the `search_teaching_events` method to the new
`search_teaching_events_indexed_by_type` as events are categorsied in the
majority of places.

Updates `events#show_category` to use new categorised events method. The amount
is set somewhat arbitrarily to 1,000 for now, we are expecting to change this
in the future to be paginated.