Support both HTML and plain text event descriptions
The current event descriptions are plain text (with line breaks) - the CRM team
are  migrating to a WYSIWYG for inputting event descriptions that will return
HTML. We need to be able to support both while they perform the migration.

Checks to see if the event description contains any HTML tags, if it does then
it runs it through `sanitize`, otherwise continue to use `simple_format`.

If we run through both `simple_format` and `sanitize` instead we end up getting
extra spacing between lines.