Rate limit transaction endpoints
The transaction endpoints are rate limited to 30 requests per minute by the
API. We want to enforce the rate limiting client-side per user/ip as well to a
maximum of 5 requests/IP/minute.

If a user encounters a rate limit, instead of returning a standard 429 and text
response we instead render a 'Too many requests' error page, similar to the
existing 404 and 500 error pages.