Sanitize metrics for better aggregation
Aggregating the metrics on path doesn't provide a good insight at present due
to the query parameters being included in the path. By removing the query
parameters we can aggregate on the underlying path/action.

Similarly, the view metrics contain the absolute path to the views directory
which makes displaying them in a table awkward. Removing the
`/path/to/app/views/` prefix leaves us with the template identifier.