Fix for :back links not re-running js in production.
In non-cdn environments, `history.back()` will result in the earlier page being
reloaded.

In our production environment, `history.back()` restores the earlier page but
does not reload the content, or re-run the JS which means the cookie acceptance
overlay continues to be shown, even though the user has accepted cookies by this
point.