Display no results message per category
When searching for events, if the user is *not* searching by a specific type
and they get results for some but not all event categories, we want to still
display the empty event categories with a 'no results' message in each one. If
they get no results for any categories, we want to display a single 'no
results' message (as per the current behaviour).

When searching for events by type, we only want to display a single 'no
results' message (as per the current behaviour).