Add a context for including fake views
Our content is in a separate repository to the main application and
is *not present* when running on CI. As a result, when attempting to
render partials form the content repo cause failures in CI. This change
adds `spec/support/views` that can be prepended to the view path via the
"prepend fake views" shared context.

As there's only one instance of this at the moment it needs to be done
on an ad hoc basis; if we get more than a few it might be worth
prepending globally.