Remove telephone field from Contact step
Based on the data we've seen so far, it appears that users at a higher
commitment will provide their phone number more often than lower commitment
people.

There is currently no value behind this offering, so we're going to remove it
from the flow, and instead, ask users to self serve once sign up is complete
(another ticket to follow on the content for the confirmation page).