Home page section updates
- Tweak mailing list CTA

Add in white space above the CTA, bound it to the container-1000 width and
left-align the paragraph.

- Update home page steps design

Make the boxes square, remove the off-white background and make the strapline
smaller. Make them a bit larger on mobile.

- New events section

Change find an event section to behave more like the quote section (text and
image). Tweak the image width on the quote section to be smaller to match new
design.

- Update featured items

New 'explore' title added. Title for each of the cards moved to below the
image.