Fix event group spacing when there are no TTT events
The Train to Teach events have a logo in the top right, which adds on a margin
to the top to push the box down below the heading. When there are no TTT events
there still needs to be a margin at the top to give space below the heading.