Simplify TalkToUs Stimulus controller
This controller previously had two targets, `chat` and `tta`. The `tta` target
was unused so could be removed and the `chat` target was used to show the chat
CTA if Javascript is enabled.

There was a bug introduced when another chat partial used the
`TalkToUsController` but didn't specify a `chat` target. Instead, this commit
refactors the controller to use the built in `element` target as the chat
element. It now also accepts a display type as DOM state, since the other use
case required a `flex` display type (and the initial use case used
`inline-block`).

Updates the test cases to simplify the stub HTML and cover the `startChat`
method.