Allow partials list in frontmatter to be rendered
The previous approach to allowing pages that are a mix of Markdown and
regular HTML - mainly used where we need calls to action mid way down a
page - was to rename the page from `.md` to `.html.erb`, and then to
render markdown pages alongside HTML. At the top of these files a
`@front_matter` instance variable is declared and set with the
appropriate metadata.

This was a flawed approach because it means that the functionality that
builds the sitemap and navigation doesn't pick up pages. This solves the
problem by reverting all templates back to markdown and allowing the
writer to create a list of partials under the `content:` key, that will
be rendered on the page. The partials can be either `.md` or
`.html.erb`, so we have complete control of the page contents.

The disadvantage of this approach is that it's more complex than a
single HTML file and we end up with a file per section plus a file per
CTA. On the plus side the syntax is straightforward and it works in a
simple manner, and the sections are very easy to reorder.