Fix JS error on guidance pages
We ran into a bug with Turbolinks imploding once you visit the a guidance page
(so the browser back button failed to work). This was caused by a JS exception
on the guidance pages, originating from some include JS. Removing the offending
JS fixes the issues.

I gave the guidance page a good click about and nothing appears to be effected
by removing this. We're also going to be removing/restyling these pages in the
near future.