Add support for Online Train to Teach Events
At the moment we have events that have the Online Event type and other events
that are marked as being `is_online` but have an associated `building` - these
are "virtual" events.

A new event type is being added, which is a Train to Teach type event that
`is_online` but has no `building`. These are to be displayed the same as
regular Train to Teach type events, except instead of having "Event has moved
online" we will show them as "Online event" (with a purple icon instead of the
current blue used for Online Event type events).

We will be able to simplify the `moved_online?` logic a little shortly as an
`is_virtual` attribute is being added to the API to help distinguish the
different states more easily.