Move stories content to a semantic layout
This is the first part of a migration away from the `content` and
`content-1000` float-based layouts and towards a simpler flexbox model.

Here we introduce the `semantic` class; this is a stopgap to target only
the pages we're transitioning while the migration is in flow. Once the
old layouts are removed we can drop the class and simplify further.

The intent is to structure the contents of `main` based on the direct
child elements. Currently there are two; `article` with `aside` and a
70/30 split and `.feature` - intended for things like index pages and
lists of events - which is full-width.