Update Event::Search to explicitly search future events only
Currently the `Event::Search` class relies on the API only returning
future-dated events (as we only cache upcoming events from the CRM).

Going forward that will change as we're displaying archived/past events in the
website. Before this change happens in the API we need to explicitly search
future events _only_ in the application.

This commit ensures that when no month or the current month is selected we are
only going to search for events after the start of the current day.