Support viewing archived online events
We want to be able to view an archive of online events (only) going back a
period of 4 months. The archive page should be identical to the event category
page, but with "Past" suffixed to the category name copy and instead of showing
the next 5 months of events it should display the previous 4 months of events.
This should display up to and including the day before today, but no events on
the current date (even if the event time has passed as they stay on the main
event category page until the following day).

This commit adds a new route `/category/:category/archive` that points to the
existing event category controller action.

The action is updated to determine if the event type supports an archive and to
configure the event search `period` to be either `past` or `future`.

The `Event::Search` module contains the majority of the updates, which stem
from a new `period` parameter that supports either `:future` (current
behaviour) or `:past` (for archived events). We use this parameter to determine
the `available_months` to generate along with capping the earliest/latest dates
to query.