Improve the experience for people without flexbox
Previously the templates placed `article` before `aside` (as per the
layout on the page) but this meant if anyone visited using a browser
that doesn't support Flexbox at all[0] the aside (often a ToC) would be
listed after the article.

This change switches them so the article is first and uses `row-reverse`
to order them article -> aside. The content is aligned to the flex-end
(the left, as we're now reversed) so if there's no aside the article
will be on the left.

[0] https://caniuse.com/flexbox