Move event category actions into EventCategoriesController
Currently we have one, reasonably complex action in the `EventsController` that
deals with showing an event category and its associated archive.

This commit pulls that action out and splits it into two new actions; `show`
and `show_archive` in a new `EventCategoriesController`.

The routes have also been updated to better match what you would expect from a
Rails app, so:

```
/event_categories/:id
/event_categories/id/archive
```

Redirects are in place to ensure any old/bookmarked traffic gets
redirected, though we can probably remove the redirects in a few days if we
observe no traffic going to the old endpoints.