Make the content pages redirect-aware
This is necessary because without it renaming content pages that are
linked to from the app will result in failed builds.

Now the renaming process can be to update the content and create a new
redirect rule and then update the app with the new path; this will
result in a short period where the link text doesn't exactly match the
content page but the build will pass.