Add FeatureTableComponent
This is a table component that will be used in the new Ways to
Train page. It consists of a strapline heading and two column table
in destop, with the first column being the heading and the second
column the associated value. In mobile this changes to a single
column layout, alternating headings and values.