Render layout according to frontmatter
If a content page specifies a `layout` in the frontmatter, we want to
render the content using this instead of the default content layout.

Initially the only supported layout will be the accordion, but we can
add to this as we go along (even though these layouts only come from
the content repo it will be more secure if we only allow a whitelist
of layouts to be used).