Handle rendering index content pages
Currently if a content page exists at `page/index.md` the request path would
need to be `/page/index`. This isn't consistent with the current story pages
that are using the `stories/story/list` template - they exist at
`my-story-into-teaching/section/index.md`, however the request paths are
`my-story-into-teaching/section`.

To accommodate this, the `PagesController` will now check if the requested
template exists before rendering the page and, if it doesn't, it will fallback
to trying `template/index`.

In addition, any direct requests to `template/index` will return a 404 so that
we don't have to deal with canonical meta tags pointing back to the `/template`
page.