Remove /my-story-into-teaching/section/index content specs
The  my story into teaching list pages are in the content repo at
`/my-story-into-eaching/section/index`, however we have Rails routes for these
that render them at `.../section`. It turns out you can still access the
`/section/index` pages directly though and they display a broken/partial
listing page.

Going forward this won't be the case as we're moving the pages to be at
`/section.md` instead of `/section/index.md` and scrapping the Rails routes.
The routes have been scrapped but the content specs are still expecting a page
at `/section/index` which is no longer the case, so excluding these from the
content specs until we can move the pages in to the content repo.