Make h2/strapline font size consistent
We seem to have slight variations in the font size of straplines; a strapline
is 36px but often overriden by the `h2` tag style to 34px/24pt. I think it
looks better at 24pt so I've tweaked them all to be consistent.

Going forward it would be nice to have a set of variables for the different
font sizes, but that's a different kettle of fish/left to a separate PR.