Support home page as markdown
Currently the home page is an ERB that relies on the `PagesController` applying
the correct layout.

This moves the current home template to `home-old` (to be removed in an
following PR) and adds a new `home` template that supports a markdown version
of the home page, which will specify the content and layout as frontmatter.