Only apply width-limiter to the page's main header
The <header> element is used for both section headings and the main
heading on the page.

Here the %content-width-limiter-shared style was being applied to all
headers on the page instead of only the outermost one (which is a direct
descendant of <main>), causing them to be rendered centrally rather than
left-aligned.