Disable Turbolinks for all jump links
We already disable Turbolinks for jump links pointing to the same page, but it
seems jump links pointing to other pages in the website aren't consistently
scrolling to the correct location (sometimes the page will scroll then an image
or something further up will load and throw the scroll position off).

Its mainly the steps to become a teacher page that this effects (linked to from
the home page).