Enable page-width content in layouts
The event `SearchComponent` has a page-width background with inner content that
is contstrained to the 1000px content width. In order to enable this the
semantic layouts have been modified slightly to include a `.container` class
that applies the 1000px content width (instead of it being applied at the
`main` tag level). The new structure is:

```
<main class="semantic">
  <div>
    ... this will be page-width ...
  </div>
  <section class="container">
    <aside></aside>
    <article class="markdown overhang"></article>
    <section class="feature"></section>
  </section>
</main>
```

I've kept the changes as minimal as possible, but there's a lot that could be
cleaned up in the `SearchComponent` CSS itself. I'll raise a ticket for this
separately.

Updates event category search form action to be consistent with find an event
page, scrolling the user to the top of the results.