Transition home page to semantic layout
Migrates the home page to a semantic layout, matching the other layouts.

Adds a new CSS component `pagewidth-bg` to display a page-wide grey/yellow
background that is used on both the home page and search for events page.

A few additional CSS classes have been added so that the new changes can
co-exist with the existing home page structure while this is rolled out. We
should be able to clean a lot of it up when the old layout is removed.