Add a disclaimer layout that allows wide tables
The cookies page has information-heavy tables that don't wrap nicely.
The new dislaimer layout aims to solve this by artificially limiting the
width of inline content (p, ul, ol) to 65% but not limiting the table.

The cookies names are now wrapped in code tags and are set to wrap
before regular text, making the table content easier to read at small
screen sizes.