Allow the hero subtitle to be a hyperlink
This was one of the ideas to increase engagement with the mailing list
signup. When a `subtitle_link` key is present in the front matter the
subtitle text is a link rather than regular text. It remains unstyled
except that on hover it's underlined (via a border to match the title
above).

When no subtitle is present the text will be displayed without any
changes.