Improve responsiveness of story cards
Update the cards to match the designs (keep the same layout on mobile/desktop).

We were going from 3 columns to 1 which, with the udpated design, left a lot of
space around the text on the cards. This updates to add a two column layout
between mobile and desktop that makes better use of the space.