Don't link to the current page in main nav
Providing two links to the same page is a bad idea, it causes
unnecessary cognitive strain[0]. This is exhibited by the site logo
linking and the first element of the navigation bar (they're next to
each other on the page) both linking to `/`.

We can avoid this by using `link_to_unless_current` and styling the text
appropriately.

[0] https://www.nngroup.com/articles/duplicate-links/