Make the home links consistent
Having multiple links with the same text but different targets on a
single page could lead to confusion. Here, the 'Home' link in the header
and footer linked to `/` and `/home` respectively. Even though they're
the same page as far as the app is concerned, it's not obvious.

They have now been standardised to the `root_path`, `/`