Remove strapline class
We currently have two ways of styling h2 headings; by `.markdown > h2` and a
`.strapline` class (legacy and for places where a wrapping div prevents the
`.markdown > h2` selector applying).

This commit removes all uses of the `.strapline` class by bubbling the `h2` up
a level where possible so that the `.markdown > h2` selector applies.

There are a couple of use cases outside the `.markdown` class, such as
`.feature` and `footer`. To cover these the heading CSS has been abstracted
into a mixin and mixed in to the particular classes.

Also adds a global right-margin on the headings when in mobile, as per the
design (apart from on the feature tables where the right margin should always
be present).