Build the homepage purple box
The 'purple box' is a large CTA that appears at the top of the homepage.
In the new design it contains a header, some text, a call to action
button and a SVG graphic.

The complexity in this compenent lies in the fact that it overhangs the
right of the purple box; this means we need to wrap the regular
container in an outer container (which includes all content *plus* the
overhanging part of the SVG).

As a result of this we can't use just go down the basic route of a pair
of nested flexbox elements; the requirement for part of one element to
overlap another requires either grid or hacks with margins.

This approach uses grid; the overhang amount is controlled vertically
and horizontally via the variables `$px-either-side-of-content` and
`$px-above-content` and the remainder of the grid should automatically
fit the content.

`$px-either-side-of-content` is also used to constrain the other
homepage content accordingly, so it lines up with the purple box.