Support magic link tokens on mailing list sign up
A magic link token can be supplied to the mailing list sign up in the form of a
query string parameter:

```
GET mailinglist/signup?magic_link_token=<token>
```

The app then calls out to the API to exchange the token for a pre-filled
mailing list sign up object.

If this is successful, then the candidate skips to the step _after_
`Authenticate`, with their information pre-filled.

If the token is not valid we redirect to the first step of the mailing list
sign up form and display an error to the user, ecouraging them to proceed
through the form using the current matchback/verification token mechanism.