Always export matchback attributes
Previously the `Authenticate` step was responsible for exporting the matchback
attributes. This made sense as that step was pre-populating the store with
them, however going forward the magic link token mechanism will also be doing
this and the logic is shifting to the `Wizard` class, so it makes sense to tack
them on to the export there instead.