Consolidate access token requests into Wizard::Base
After adding the logic to exchange magic link tokens for candidate information
into the `Wizard::Base` subclasses we now have similar logic in the
`Authenticate` step subclasses.

It makes sense to approach this consistently and drop the custom `Authenticate`
step subclasses in favour of having the token exchange logic in the
`Wizard::Base` subclasses.

There is also a small change in functionality in this commit; previously the
store was purged before issuing a verification code only if a successful
authentication had happened in this candidates sign up journey (so if they
matched back then went to start the process again). Now we always purge the
store prior to issuing a new verification code.

The main reason for this is to avoid potentially complicated sign up scenarios
now that users can also authenticate with a magic link.

The only detrimental effect of this change is if a candidate goes through the
flow, then back to the first step and _doesn't_ match back; they will need to
re-enter their details instead of having them pre-filled.

This feels like enough of an edge case that its acceptable, given how much
simpler and robust the code is if we just clear everything out and treat it as
a new sign up (if you go back and change your name/email then the liklihood is
its a new/different person signing up anyway).